refactor text selected
