+++ apps/web/client/src/app/project/[id]/_components/editor-bar/dropdowns/height.tsx
@@ -5,8 +5,8 @@ import { DropdownMenu, DropdownMenuContent, DropdownMenuTrigger } from '@onlook/
 import { Icons } from '@onlook/ui/icons';
 import { LayoutMode } from '@onlook/utility';
 import { useDimensionControl } from '../hooks/use-dimension-control';
+import { HoverOnlyTooltip } from '../hover-tooltip';
 import { InputDropdown } from '../inputs/input-dropdown';
-import { HoverOnlyTooltip } from '../HoverOnlyTooltip';
 
 export const Height = () => {
     const { dimensionState, handleDimensionChange, handleUnitChange, handleLayoutChange } =
@@ -17,22 +17,22 @@ export const Height = () => {
             <HoverOnlyTooltip content="Height" side="bottom" className="mt-1" hideArrow>
                 <DropdownMenuTrigger asChild>
                     <Button
-                            variant="ghost"
-                            size="toolbar"
-                            className="text-muted-foreground border-border/0 hover:bg-background-tertiary/20 hover:border-border data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border flex cursor-pointer items-center gap-1 border hover:border hover:text-white focus:outline-none focus-visible:ring-0 focus-visible:ring-offset-0 focus-visible:outline-none active:border-0 data-[state=open]:border data-[state=open]:text-white"
-                        >
-                            <Icons.Height className="h-4 min-h-4 w-4 min-w-4" />
-                            {(dimensionState.height.unit === 'px'
-                                ? dimensionState.height.num !== undefined
-                                : (dimensionState.height.value && dimensionState.height.value !== "auto")
-                            ) && (
+                        variant="ghost"
+                        size="toolbar"
+                        className="text-muted-foreground border-border/0 hover:bg-background-tertiary/20 hover:border-border data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border flex cursor-pointer items-center gap-1 border hover:border hover:text-white focus:outline-none focus-visible:ring-0 focus-visible:ring-offset-0 focus-visible:outline-none active:border-0 data-[state=open]:border data-[state=open]:text-white"
+                    >
+                        <Icons.Height className="h-4 min-h-4 w-4 min-w-4" />
+                        {(dimensionState.height.unit === 'px'
+                            ? dimensionState.height.num !== undefined
+                            : (dimensionState.height.value && dimensionState.height.value !== "auto")
+                        ) && (
                                 <span className="text-small">
                                     {dimensionState.height.unit === 'px'
                                         ? Math.round(dimensionState.height.num ?? 0)
                                         : dimensionState.height.value}
                                 </span>
                             )}
-                        </Button>
+                    </Button>
                 </DropdownMenuTrigger>
             </HoverOnlyTooltip>
             <DropdownMenuContent
