+++ apps/web/client/src/app/project/[id]/_components/editor-bar/dropdowns/padding.tsx
@@ -7,12 +7,11 @@ import {
     DropdownMenuTrigger,
 } from "@onlook/ui/dropdown-menu";
 import { Icons } from "@onlook/ui/icons";
-import { Tooltip, TooltipContent, TooltipTrigger } from "@onlook/ui/tooltip";
 import { useState } from "react";
 import { useBoxControl } from "../hooks/use-box-control";
+import { HoverOnlyTooltip } from "../hover-tooltip";
 import { InputRange } from "../inputs/input-range";
 import { SpacingInputs } from "../inputs/spacing-inputs";
-import { HoverOnlyTooltip } from "../HoverOnlyTooltip";
 
 export const Padding = () => {
     const [activeTab, setActiveTab] = useState('all');
@@ -23,18 +22,18 @@ export const Padding = () => {
             <HoverOnlyTooltip content="Padding" side="bottom" className="mt-1" hideArrow>
                 <DropdownMenuTrigger asChild>
                     <Button
-                            variant="ghost"
-                            size="toolbar"
-                            className="text-muted-foreground border-border/0 hover:bg-background-tertiary/20 hover:border-border data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border gap-1 flex cursor-pointer items-center border hover:border hover:text-white focus:outline-none focus-visible:ring-0 focus-visible:ring-offset-0 focus-visible:outline-none active:border-0 data-[state=open]:border data-[state=open]:text-white"
-                        >
-                            <Icons.Padding className="h-4 min-h-4 w-4 min-w-4" />
-                            {boxState.padding.unit === 'px' && typeof boxState.padding.num === 'number' && boxState.padding.num !== 0 ? (
-                                <span className="text-small">{boxState.padding.num}</span>
-                            ) : null}
-                            {boxState.padding.unit !== 'px' && boxState.padding.value ? (
-                                <span className="text-small">{boxState.padding.value}</span>
-                            ) : null}
-                        </Button>
+                        variant="ghost"
+                        size="toolbar"
+                        className="text-muted-foreground border-border/0 hover:bg-background-tertiary/20 hover:border-border data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border gap-1 flex cursor-pointer items-center border hover:border hover:text-white focus:outline-none focus-visible:ring-0 focus-visible:ring-offset-0 focus-visible:outline-none active:border-0 data-[state=open]:border data-[state=open]:text-white"
+                    >
+                        <Icons.Padding className="h-4 min-h-4 w-4 min-w-4" />
+                        {boxState.padding.unit === 'px' && typeof boxState.padding.num === 'number' && boxState.padding.num !== 0 ? (
+                            <span className="text-small">{boxState.padding.num}</span>
+                        ) : null}
+                        {boxState.padding.unit !== 'px' && boxState.padding.value ? (
+                            <span className="text-small">{boxState.padding.value}</span>
+                        ) : null}
+                    </Button>
                 </DropdownMenuTrigger>
             </HoverOnlyTooltip>
             <DropdownMenuContent align="start" className="w-[280px] mt-1 p-3 rounded-lg">
@@ -42,8 +41,8 @@ export const Padding = () => {
                     <button
                         onClick={() => setActiveTab('all')}
                         className={`flex-1 text-sm px-4 py-1.5 rounded-md transition-colors cursor-pointer ${activeTab === 'all'
-                                ? 'text-white bg-background-tertiary/20'
-                                : 'text-muted-foreground hover:bg-background-tertiary/10'
+                            ? 'text-white bg-background-tertiary/20'
+                            : 'text-muted-foreground hover:bg-background-tertiary/10'
                             }`}
                     >
                         All sides
@@ -51,8 +50,8 @@ export const Padding = () => {
                     <button
                         onClick={() => setActiveTab('individual')}
                         className={`flex-1 text-sm px-4 py-1.5 rounded-md transition-colors cursor-pointer ${activeTab === 'individual'
-                                ? 'text-white bg-background-tertiary/20'
-                                : 'text-muted-foreground hover:bg-background-tertiary/10'
+                            ? 'text-white bg-background-tertiary/20'
+                            : 'text-muted-foreground hover:bg-background-tertiary/10'
                             }`}
                     >
                         Individual
