+++ apps/web/client/src/app/project/[id]/_components/editor-bar/dropdowns/radius.tsx
@@ -9,7 +9,7 @@ import {
 import { Icons } from "@onlook/ui/icons";
 import { useState } from "react";
 import { useBoxControl } from "../hooks/use-box-control";
-import { HoverOnlyTooltip } from "../HoverOnlyTooltip";
+import { HoverOnlyTooltip } from "../hover-tooltip";
 import { InputRange } from "../inputs/input-range";
 import { SpacingInputs } from "../inputs/spacing-inputs";
