+++ apps/web/client/src/app/project/[id]/_components/editor-bar/hover-tooltip.tsx
