+++ apps/web/client/src/app/project/[id]/_components/editor-bar/text-selected/font-size.tsx
@@ -1,0 +1,105 @@
+'use client';
+
+import { Button } from '@onlook/ui/button';
+import { Icons } from '@onlook/ui/icons';
+import { Popover, PopoverContent, PopoverTrigger } from '@onlook/ui/popover';
+import { Tooltip, TooltipContent, TooltipTrigger } from '@onlook/ui/tooltip';
+import { memo, useRef, useState } from 'react';
+
+const FONT_SIZES = [12, 14, 16, 18, 20, 24, 30, 36, 48, 60, 72, 96];
+
+export const FontSizeSelector = memo(
+    ({
+        fontSize,
+        handleFontSizeChange,
+    }: {
+        fontSize: number;
+        handleFontSizeChange: (size: number) => void;
+    }) => {
+        const [open, setOpen] = useState(false);
+        const inputRef = useRef<HTMLInputElement>(null);
+
+        const adjustFontSize = (amount: number) => {
+            handleFontSizeChange(Math.max(1, fontSize + amount));
+        };
+
+        const handleInputClick = () => {
+            setOpen(true);
+            // Use setTimeout to ensure the input is focused after the popover opens
+            setTimeout(() => {
+                inputRef.current?.focus();
+                inputRef.current?.select();
+            }, 0);
+        };
+
+        return (
+            <Popover open={open} onOpenChange={setOpen}>
+                <Tooltip>
+                    <div>
+                        <TooltipTrigger asChild>
+                            <div className="flex items-center gap-0.5">
+                                <Button
+                                    variant="ghost"
+                                    size="toolbar"
+                                    onClick={() => adjustFontSize(-1)}
+                                    className="border-border/0 hover:bg-background-tertiary/20 hover:border-border text-muted-foreground data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border h-8 w-8 cursor-pointer rounded-lg border px-2 hover:border hover:text-white data-[state=open]:border data-[state=open]:text-white"
+                                >
+                                    <Icons.Minus className="h-4 w-4" />
+                                </Button>
+                                <PopoverTrigger asChild>
+                                    <input
+                                        ref={inputRef}
+                                        type="number"
+                                        value={fontSize}
+                                        onChange={(e) => {
+                                            const value = parseInt(e.target.value);
+                                            if (!isNaN(value) && value > 0) {
+                                                handleFontSizeChange(value);
+                                            }
+                                        }}
+                                        onClick={handleInputClick}
+                                        className="border-border/0 text-muted-foreground hover:bg-background-tertiary/20 hover:border-border data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border focus:bg-background-tertiary/20 focus:ring-border h-8 max-w-[40px] min-w-[40px] [appearance:textfield] rounded-lg border px-1 text-center text-sm hover:border hover:text-white focus:ring-1 focus:outline-none data-[state=open]:border data-[state=open]:text-white [&::-webkit-inner-spin-button]:appearance-none [&::-webkit-outer-spin-button]:appearance-none"
+                                    />
+                                </PopoverTrigger>
+
+                                <Button
+                                    variant="ghost"
+                                    size="toolbar"
+                                    onClick={() => adjustFontSize(1)}
+                                    className="border-border/0 hover:bg-background-tertiary/20 hover:border-border text-muted-foreground data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border h-8 w-8 cursor-pointer rounded-lg border px-2 hover:border hover:text-white data-[state=open]:border data-[state=open]:text-white"
+                                >
+                                    <Icons.Plus className="h-4 w-4" />
+                                </Button>
+                            </div>
+                        </TooltipTrigger>
+                        <TooltipContent side="bottom" className="mt-1" hideArrow>
+                            Font Size
+                        </TooltipContent>
+                    </div>
+                </Tooltip>
+                <PopoverContent
+                    align="center"
+                    className="mt-1 w-[48px] min-w-[48px] rounded-lg p-1"
+                >
+                    <div className="grid grid-cols-1 gap-1">
+                        {FONT_SIZES.map((size) => (
+                            <button
+                                key={size}
+                                onClick={() => {
+                                    handleFontSizeChange(size);
+                                    setOpen(false);
+                                }}
+                                className={`cursor-pointer text-muted-foreground data-[highlighted]:bg-background-tertiary/10 border-border/0 data-[highlighted]:border-border justify-center rounded-md border px-2 py-1 text-sm data-[highlighted]:text-white ${size === fontSize
+                                    ? 'bg-background-tertiary/20 border-border border text-white'
+                                    : ''
+                                    }`}
+                            >
+                                {size}
+                            </button>
+                        ))}
+                    </div>
+                </PopoverContent>
+            </Popover>
+        );
+    },
+);
