+++ apps/web/client/src/app/project/[id]/_components/editor-bar/text-selected/text-align.tsx
@@ -1,0 +1,78 @@
+'use client';
+
+import { Button } from '@onlook/ui/button';
+import {
+    DropdownMenu,
+    DropdownMenuContent,
+    DropdownMenuItem,
+    DropdownMenuTrigger,
+} from '@onlook/ui/dropdown-menu';
+import { Icons } from '@onlook/ui/icons';
+import { Tooltip, TooltipContent, TooltipTrigger } from '@onlook/ui/tooltip';
+import { memo } from 'react';
+import { type TextAlign } from '../hooks/use-text-control';
+
+export const TextAlignSelector = memo(
+    ({
+        textAlign,
+        handleTextAlignChange,
+    }: {
+        textAlign: TextAlign;
+        handleTextAlignChange: (align: TextAlign) => void;
+    }) => (
+        <DropdownMenu>
+            <Tooltip>
+                <div>
+                    <TooltipTrigger asChild>
+                        <DropdownMenuTrigger asChild>
+                            <Button
+                                variant="ghost"
+                                size="toolbar"
+                                className="text-white border-border/0 hover:bg-background-tertiary/20 hover:border-border data-[state=open]:bg-background-tertiary/20 data-[state=open]:border-border flex max-w-9 min-w-9 cursor-pointer items-center justify-center gap-2 rounded-lg border px-2 hover:border hover:text-white focus:outline-none focus-visible:ring-0 focus-visible:ring-offset-0 focus-visible:outline-none active:border-0 data-[state=open]:border data-[state=open]:text-white"
+                            >
+                                {(() => {
+                                    switch (textAlign) {
+                                        case 'center':
+                                            return <Icons.TextAlignCenter className="h-4 w-4" />;
+                                        case 'right':
+                                            return <Icons.TextAlignRight className="h-4 w-4" />;
+                                        case 'justify':
+                                            return <Icons.TextAlignJustified className="h-4 w-4" />;
+                                        case 'left':
+                                        default:
+                                            return <Icons.TextAlignLeft className="h-4 w-4" />;
+                                    }
+                                })()}
+                            </Button>
+                        </DropdownMenuTrigger>
+                    </TooltipTrigger>
+                    <TooltipContent side="bottom" className="mt-1" hideArrow>
+                        Text Align
+                    </TooltipContent>
+                </div>
+            </Tooltip>
+            <DropdownMenuContent
+                align="center"
+                className="mt-1 flex min-w-fit gap-1 rounded-lg p-1"
+            >
+                {[
+                    { value: 'left' as TextAlign, icon: Icons.TextAlignLeft },
+                    { value: 'center' as TextAlign, icon: Icons.TextAlignCenter },
+                    { value: 'right' as TextAlign, icon: Icons.TextAlignRight },
+                    { value: 'justify' as TextAlign, icon: Icons.TextAlignJustified },
+                ].map(({ value, icon: Icon }) => (
+                    <DropdownMenuItem
+                        key={value}
+                        onClick={() => handleTextAlignChange(value)}
+                        className={`text-muted-foreground data-[highlighted]:bg-background-tertiary/10 border-border/0 data-[highlighted]:border-border rounded-md border px-2 py-1.5 data-[highlighted]:text-foreground cursor-pointer transition-colors duration-150 hover:bg-background-tertiary/20 hover:text-foreground ${textAlign === value
+                            ? 'bg-background-tertiary/20 border-border border text-white'
+                            : ''
+                            }`}
+                    >
+                        <Icon className="h-4 w-4" />
+                    </DropdownMenuItem>
+                ))}
+            </DropdownMenuContent>
+        </DropdownMenu>
+    ),
+);
